Namespace specifically for handling errors
A new error global flag has been added. This is modified through select
functions which modify the state. This relies on the computations in the tree being non-lazy,
so functions like map which would return lazy computations are forced to execute. Since the
results of the parse will be immediately used in the system down the line, these souldn't
have a performance impact.