Add without-errors->> macro
This macro is based on the some->> macro the clojure.core, but adds an additional check to each
step. In addition to checking that the previous result not being nil, it also checks that an error
has not been raised in the system, through raise-error!. This ensures that we can thread the core
compiler functions in a clean way without having to deal with nested if statements down the line
when more functions beyond parse need to be called.